Project completed

+++ controllers/contact-controller.js
@@ -4,15 +4,16 @@ const Contact = require("../models/contact-model")
 
 //@desc Get All Contact
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find()
+    const contacts = await Contact.find({user_id: req.user.id })
+    console.log(req.user.id)
     res.status(200).json(contacts)
 });
 
 //@desc Create Contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("Request Body is: ", req.body)
     const {name, email, phone} = req.body;
@@ -23,7 +24,8 @@ const createContact = asyncHandler(async (req, res) => {
     const contacts = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     });
     console.log(contacts)
     res.status(201).json(contacts);
@@ -35,12 +37,16 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc delete Contact
 //@route DELETE /api/contacts:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res)=> {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact does not exist!")
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("User don't have permission to delete other user contacts")
     }
     await Contact.findByIdAndDelete(req.params.id)
     res.status(200).json(contact)
@@ -52,26 +58,36 @@ const deleteContact = asyncHandler(async (req, res)=> {
 
 //@desc Update Contact
 //@route PUT /api/contacts:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact does not exist!")
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("User don't have permission to update other user contacts")
+    }
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {new: true})
     res.status(200).json(updatedContact)
 });
 
+
 //@desc Get Single Contact
 //@route GET /api/contacts:id
-//@access public
+//@access private
 const getSingleContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact){
         res.status(404)
         throw new Error("Contact Not Found");
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("User don't have permission to view other user contacts")
+    }
+
     res.status(200).json(contact)
 });
